Migrate tests for contacts manager to correct test class
